Refactoring Modal
