+++ superformula-frontend/src/components/Modal/Modal.module.scss
@@ -4,6 +4,6 @@
     left: 0;
     right: 0;
     bottom: 0;
-    background-color: rgba(0, 0, 0, 0.4);
+    background-color: var(--overlay);
     z-index: 1;
 }
