+++ superformula-frontend/src/components/Modal/Modal.tsx
@@ -1,22 +1,23 @@
-import React, { ReactElement } from 'react'
+import React, { ReactElement, ReactNode } from 'react'
 import ReactDOM from 'react-dom'
-import { ModalProps } from './Modal.types'
 import globalStyles from '../../styles/themes.module.scss'
 import styles from './Modal.module.scss'
 
-function Modal({
-    children,
-    backgroundClassName,
-    className,
-    isShowing,
-}: ModalProps): ReactElement | null {
+type ModalProps = {
+    children: ReactNode
+    backgroundClassName?: string
+    className?: string
+}
+
+function Modal({ children, backgroundClassName, className }: ModalProps): ReactElement | null {
     const getModalWrapper = () => (
         <div className={`${globalStyles.defaultTheme} ${styles.background} ${backgroundClassName}`}>
             <div className={className}>{children}</div>
         </div>
     )
 
-    return isShowing ? ReactDOM.createPortal(getModalWrapper(), document.body) : null
+    return ReactDOM.createPortal(getModalWrapper(), document.body)
 }
 
 export default Modal
+export type { ModalProps }
