--- superformula-frontend/src/components/Modal/Modal.types.ts
@@ -1,10 +1,0 @@
-import { ReactNode } from 'react'
-
-type ModalProps = {
-    children: ReactNode
-    isShowing: boolean
-    backgroundClassName?: string
-    className?: string
-}
-
-export type { ModalProps }
