+++ superformula-frontend/src/components/Modal/index.tsx
@@ -1,0 +1,1 @@
+export * from './Modal'
