+++ superformula-frontend/src/styles/themes.module.scss
@@ -9,6 +9,8 @@
 
     --border-color: rgba(0, 0, 0, 0.1);
 
+    --overlay: rgba(0, 0, 0, 0.4);
+
     --elevation-z1: 0px 4px 10px rgba(0, 0, 0, 0.1);
 
     --typography-h1: 6rem;
