+++ superformula-frontend/src/userManagement/components/EditModal/EditModal.types.ts
@@ -2,7 +2,6 @@ import { User } from '../../models/User'
 
 type EditModalProps = {
     user: User
-    isShowing: boolean
     onSave: () => void
     onClose: () => void
 }
